requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
mysqlclient==2.1.0
openpyxl==3.0.9
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
python-dateutil==2.8.2
django-background-tasks==1.2.5
 pillow==9.1.0
 boto3==1.21.32
 botocore==1.24.32
 jmespath==1.0.0
 s3transfer==0.5.2